+++ CHANGELOG.md
@@ -27,6 +27,8 @@ for now removed features.
 
 ### Fixed
 
+- **alert**: is now declared in wcs-angular module
+- **alert-drawer**: is now declared in wcs-angular module
 - **card-header**: is now declared in wcs-angular module
 - **card-content**: is now declared in wcs-angular module
 - **card-footer**: is now declared in wcs-angular module